requirements.txt
matplotlib>=3.8.4
numpy>=1.25.2
pandas>=2.2.2
PyYAML>=6.0.2
requests>=2.32.5
scikit-learn>=1.7.1
seaborn>=0.13.2
xgboost>=2.1.0
prefect>=2.0.0
